requirements.txt
# Core functionality
feedparser==6.0.11
requests==2.31.0
beautifulsoup4==4.12.3

# NLP / Processing
openai==1.12.0             # If using OpenAI models for summarization
tqdm==4.66.4               # Progress bars for loops

# Vector storage / semantic search
faiss-cpu==1.8.0           # or use pinecone-client / weaviate-client if using those services
sentence-transformers==2.6.1

# Scheduling (optional)
APScheduler==3.10.4

# Logging and utility
loguru==0.7.2

# Database support
pymongo==4.6.3             # For MongoDB
SQLAlchemy==2.0.30         # For SQL-based storage (PostgreSQL, SQLite, etc.)

# Scraping with JS support
playwright==1.44.0

# Graph database
neo4j==5.17.0

# Environment and HTTP
python-dotenv==1.0.0
httpx==0.27.2

# LangChain and ML
langchain-huggingface

# Testing
pytest
pytest-dotenv

# Deep learning
torch==2.2.2
numpy==1.26.4
transformers==4.52.4

# Parallel processing dependencies
aiohttp==3.9.1